fixup! fix(accounts): fix update of the accounts limits
